Add CustomTextFieldDelegate in CustomTextField
CustomTextFieldDelegate(이전에는 ValidateTextFieldDelegate였습니다) 를 이제는
CustomTextField에서 초기화할 때 같이 반영해주는 방식으로 하였습니다.
그 외에도 텍스트 필드에서 텍스트를 지울 때 오류가 나는 부분 수정하였습니다.
